Changes has been done

+++ app/Http/Controllers/Lead/LeadController.php
@@ -39,35 +39,52 @@ class LeadController extends Controller
             $lead = Lead::create($request->all());
             return response()->json(['status' => true, 'message' => trans('messages.lead.lead_created')]);
         }catch (\Exception $e) {
+            \Log::info($e->getMessage().' '.$e->getFile().' '.$e->getLine());
             return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
         }
     }
 
-    public function edit($lang, Lead $lead)
+    public function edit(Request $request)
     {
-        $campaigns = Campaign::orderBy('id','desc')->get();
-        $htmlView = view('lead.edit', compact('campaigns','lead'))->render();
-        return response()->json(['success' => true, 'htmlView' => $htmlView]);
+        try {
+            $lead = Lead::findOrFail($request->lead);
+            $campaigns = Campaign::orderBy('id','desc')->get();
+            $htmlView = view('lead.edit', compact('campaigns','lead'))->render();
+            return response()->json(['success' => true, 'htmlView' => $htmlView]);
+        } catch (\Exception $e) {
+            \Log::info($e->getMessage().' '.$e->getFile().' '.$e->getLine());
+            return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
+        }
     }
 
-    public function update($lang,UpdateRequest $request, Lead $lead)
+    public function update(UpdateRequest $request)
     {
         try{
-            $lead->update($request->all());
-            return response()->json(['status' => true, 'message' => trans('messages.lead.lead_updated')]);
+            if($request->ajax()){
+                $input = $request->validated();
+                $lead = Lead::findOrFail($request->lead);
+                $lead->update($input);
+                return response()->json(['status' => true, 'message' => trans('messages.lead.lead_updated')]);
+            }
+            return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
         }catch (\Exception $e) {
+            \Log::info($e->getMessage().' '.$e->getFile().' '.$e->getLine());
             return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
         }
     }
 
-    public function destroy($lang,Lead $lead)
+    public function destroy(Request $request)
     {
         abort_if(Gate::denies('leads_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
-            $lead->delete();
-            return response()->json(['status' => true, 'message' => trans('messages.lead.lead_updated')]);
-        }
-        catch (\Exception $e) {
+            if($request->ajax()){
+                $lead = Lead::whereId($request->lead);
+                $lead->delete();
+                return response()->json(['status' => true, 'message' => trans('messages.lead.lead_updated')]);
+            }
+            return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
+        } catch (\Exception $e) {
+            \Log::info($e->getMessage().' '.$e->getFile().' '.$e->getLine());
             return response()->json(['status' => false, 'message' => trans('messages.error_message')], 500);
         }
     }
